Almost there !
So, the main app is now working AND does not crash hard.
Boy, playing Process in C# can lead to the most hideous
crashes in Windows, that can lead to almost freezing the
machine.

So, now, when AdbHandler exits the tree, I kill any "adb"
process.
This is a bit "overkill", but this works for the moment.

Signed-off-by: Voyage <voyage@miouyouyou.fr>